Move git clone as function for git provider
